+++ src/main/java/com/zipte/platform/server/adapter/in/web/EstateApi.java
@@ -3,9 +3,8 @@ package com.zipte.platform.server.adapter.in.web;
 import com.zipte.platform.core.response.ApiResponse;
 import com.zipte.platform.core.response.pageable.PageRequest;
 import com.zipte.platform.core.response.pageable.PageResponse;
-import com.zipte.platform.server.adapter.in.web.dto.EstateDetailResponse;
-import com.zipte.platform.server.adapter.in.web.dto.EstateListResponse;
-import com.zipte.platform.server.adapter.in.web.dto.EstateLocationRequest;
+import com.zipte.platform.server.adapter.in.web.dto.response.EstateDetailResponse;
+import com.zipte.platform.server.adapter.in.web.dto.response.EstateListResponse;
 import com.zipte.platform.server.application.in.estate.GetEstateUseCase;
 import com.zipte.platform.server.domain.estate.Estate;
 import jakarta.persistence.EntityNotFoundException;
