+++ src/main/java/com/zipte/platform/server/adapter/in/web/PropertyApi.java
@@ -3,9 +3,9 @@ package com.zipte.platform.server.adapter.in.web;
 import com.zipte.platform.core.response.ApiResponse;
 import com.zipte.platform.core.response.pageable.PageRequest;
 import com.zipte.platform.core.response.pageable.PageResponse;
-import com.zipte.platform.server.adapter.in.web.dto.PropertyDetailResponse;
-import com.zipte.platform.server.adapter.in.web.dto.PropertyListResponse;
-import com.zipte.platform.server.adapter.in.web.dto.PropertyRequest;
+import com.zipte.platform.server.adapter.in.web.dto.response.PropertyDetailResponse;
+import com.zipte.platform.server.adapter.in.web.dto.response.PropertyListResponse;
+import com.zipte.platform.server.adapter.in.web.dto.request.PropertyRequest;
 import com.zipte.platform.server.application.in.property.*;
 import com.zipte.platform.server.domain.property.Property;
 import jakarta.validation.Valid;
