+++ src/main/java/com/zipte/platform/server/adapter/in/web/dto/response/PropertySnippetResponse.java
@@ -1,4 +1,4 @@
-package com.zipte.platform.server.adapter.in.web.dto;
+package com.zipte.platform.server.adapter.in.web.dto.response;
 
 import com.zipte.platform.server.domain.property.PropertyAddress;
 import com.zipte.platform.server.domain.property.PropertySnippet;
