+++ src/main/java/com/zipte/platform/server/adapter/in/web/dto/response/PropertyStatisticsResponse.java
@@ -1,4 +1,4 @@
-package com.zipte.platform.server.adapter.in.web.dto;
+package com.zipte.platform.server.adapter.in.web.dto.response;
 
 import com.zipte.platform.server.domain.property.PropertyStatistic;
 import lombok.Builder;
